requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
email-validator==2.1.0.post1
werkzeug==3.0.1
openai>=1.60.0
httpx==0.27.2
python-dotenv==1.0.0
gunicorn==21.2.0
